# auth-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API documentation for auth",
        "title": "auth API",
        "contact": {
            "email": "radhian.amri@gmail.com"
        },
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/claims": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "attempts extract user claims",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token with format 'Bearer {jwt_token}'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "attempts to login user",
                "parameters": [
                    {
                        "description": "Body request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "attempts to create new user",
                "parameters": [
                    {
                        "description": "Body request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.UserRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/swagdocs": {
            "get": {
                "tags": [
                    "Swagger"
                ],
                "summary": "get Swagger Documentation"
            }
        }
    },
    "definitions": {
        "model.UserLogin": {
            "type": "object",
            "required": [
                "password",
                "phone"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "model.UserRegister": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}